+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+* Fix unlock edit form in `edit-form` route and `lock-route`
+* `flexberry-objectlistview-route` mixin:
+    * Fix set current controller during the transition on edit form
 
 ### Added
 * Blueprints: